+++ Changelog.md
@@ -5,7 +5,7 @@ Language Features:
 
 Compiler Features:
  * Metadata: Added support for IPFS hashes of large files that need to be split in multiple chunks.
-
+ * Commandline Interface: Enable output of storage layout with `--storage-layout`.
 
 Bugfixes:
  * Inline Assembly: Fix internal error when accessing invalid constant variables.